+++ app/controllers/facilities_management/rm6378/home_controller.rb
@@ -1,0 +1,11 @@
+module FacilitiesManagement
+  module RM6378
+    class HomeController < FacilitiesManagement::FrameworkController
+      include SharedPagesConcern
+
+      skip_before_action :authenticate_user!, :authorize_user, :redirect_to_buyer_detail
+
+      def index; end
+    end
+  end
+end