+++ app/models/concerns/working_days.rb
@@ -3,14 +3,15 @@ module WorkingDays
     first_working_day = start_date
     first_working_day = (first_working_day + 1.day).beginning_of_day while weekend?(first_working_day) || bank_holiday?(first_working_day)
 
-    end_date = (first_working_day + number.days)
-    dates = (first_working_day..end_date).to_a
+    end_date = first_working_day
+    number_of_days = 0
 
-    dates.each do |date|
-      dates << (dates.last + 1.day) if weekend?(date) || bank_holiday?(date)
+    while number_of_days < number
+      end_date += 1.day
+      number_of_days += 1 unless weekend?(end_date) || bank_holiday?(end_date)
     end
 
-    first_working_day == start_date ? dates.last : dates.last - (first_working_day.in_time_zone('London').utc_offset / 3600).hour
+    first_working_day == start_date ? end_date : end_date - (first_working_day.in_time_zone('London').utc_offset / 3600).hour
   end
 
   def self.weekend?(date)
@@ -18,15 +19,6 @@ module WorkingDays
   end
 
   def self.bank_holiday?(date)
-    date = date.to_date
-    uk_bank_holidays.include? date
-  end
-
-  def self.uk_bank_holidays
-    @uk_bank_holidays ||= bank_holiday_json.map { |country| country[1]['events'].map { |date| DateTime.parse(date['date']).in_time_zone('London') } }.flatten.uniq
-  end
-
-  def self.bank_holiday_json
-    JSON.parse(Net::HTTP.get(URI('https://www.gov.uk/bank-holidays.json')))
+    BankHoliday.bank_holiday?(date)
   end
 end