+++ app/models/supplier/framework/address.rb
@@ -1,0 +1,7 @@
+class Supplier < ApplicationRecord
+  class Framework < ApplicationRecord
+    class Address < ApplicationRecord
+      belongs_to :supplier_framework, inverse_of: :address, class_name: 'Supplier::Framework'
+    end
+  end
+end