+++ features/step_definitions/hooks.rb
@@ -50,17 +50,39 @@ After('@mobile') do
   resize_window_to_pc
 end
 
+Before do
+  if @framework == 'RM3830' && FacilitiesManagement::RM3830::SupplierDetail.none?
+    Rake::Task['db:rm3830:fm_supplier_data'].reenable
+    Rake::Task['db:rm3830:add_supplier_rate_cards'].reenable
+
+    Rake::Task['db:rm3830:fm_supplier_data'].invoke
+    Rake::Task['db:rm3830:add_supplier_rate_cards'].invoke
+  elsif @framework == 'RM6232' && FacilitiesManagement::RM6232::Supplier.none?
+    Rake::Task['db:rm6232:import_suppliers'].reenable
+
+    Rake::Task['db:rm6232:import_suppliers'].invoke
+  end
+end
+
 After do
   DatabaseCleaner.clean
   if Framework.none?
-    Rake::Task['db:fm_frameworks'].reenable
-    Rake::Task['db:rm3830:fm_supplier_data'].reenable
-    Rake::Task['db:rm3830:add_supplier_rate_cards'].reenable
-    Rake::Task['db:rm6232:import_suppliers'].reenable
+    Rake::Task['db:frameworks'].reenable
+    Rake::Task['db:make_rm6232_live'].reenable
 
-    Rake::Task['db:fm_frameworks'].invoke
-    Rake::Task['db:rm3830:fm_supplier_data'].invoke
-    Rake::Task['db:rm3830:add_supplier_rate_cards'].invoke
-    Rake::Task['db:rm6232:import_suppliers'].invoke
+    Rake::Task['db:frameworks'].invoke
+    Rake::Task['db:make_rm6232_live'].invoke
+
+    if @framework == 'RM3830'
+      Rake::Task['db:rm3830:fm_supplier_data'].reenable
+      Rake::Task['db:rm3830:add_supplier_rate_cards'].reenable
+
+      Rake::Task['db:rm3830:fm_supplier_data'].invoke
+      Rake::Task['db:rm3830:add_supplier_rate_cards'].invoke
+    elsif @framework == 'RM6232'
+      Rake::Task['db:rm6232:import_suppliers'].reenable
+
+      Rake::Task['db:rm6232:import_suppliers'].invoke
+    end
   end
 end