+++ spec/controllers/facilities_management/admin/home_controller_spec.rb
@@ -4,27 +4,27 @@ RSpec.describe FacilitiesManagement::Admin::HomeController do
   let(:default_params) { { service: 'facilities_management/admin' } }
 
   describe 'GET framework' do
-    it 'redirects to the RM6232 admin home page' do
+    it 'redirects to the RM6378 admin home page' do
       get :framework
 
-      expect(response).to redirect_to facilities_management_admin_index_path('RM6232')
+      expect(response).to redirect_to facilities_management_admin_index_path('RM6378')
     end
 
-    context 'when RM3830 is live' do
-      include_context 'and RM3830 is live'
-      include_context 'and RM6232 is live in the future'
+    context 'when RM6232 is live' do
+      include_context 'and RM6232 is live'
+      include_context 'and RM6378 is live in the future'
 
-      it 'redirects to the RM3830 admin home page' do
+      it 'redirects to the RM6232 admin home page' do
         get :framework
 
-        expect(response).to redirect_to facilities_management_rm3830_admin_path
+        expect(response).to redirect_to facilities_management_rm6232_admin_path
       end
     end
   end
 
   describe 'GET index' do
-    context 'when RM6232 is live' do
-      context 'and the framework is not RM3830 or RM6232' do
+    context 'when RM6378 is live' do
+      context 'and the framework is not RM6232 or RM6378' do
         it 'renders the unrecognised framework page with the right http status' do
           get :index, params: { framework: 'RM3826' }
 
@@ -35,6 +35,14 @@ RSpec.describe FacilitiesManagement::Admin::HomeController do
 
       # This is because in practice, the rails router will have already used the correct framework controller,
       # therefore, this test is just to make sure that the UnrecognisedLiveFrameworkError is not invoked
+      context 'and the framework is RM6378' do
+        it 'raises the MissingExactTemplate error' do
+          expect do
+            get :index, params: { framework: 'RM6378' }
+          end.to raise_error(ActionController::MissingExactTemplate)
+        end
+      end
+
       context 'and the framework is RM6232' do
         it 'raises the MissingExactTemplate error' do
           expect do
@@ -42,20 +50,12 @@ RSpec.describe FacilitiesManagement::Admin::HomeController do
           end.to raise_error(ActionController::MissingExactTemplate)
         end
       end
-
-      context 'and the framework is RM3830' do
-        it 'raises the MissingExactTemplate error' do
-          expect do
-            get :index, params: { framework: 'RM3830' }
-          end.to raise_error(ActionController::MissingExactTemplate)
-        end
-      end
     end
 
-    context 'when RM6232 is not live' do
-      include_context 'and RM6232 is live in the future'
+    context 'when RM6378 is not live' do
+      include_context 'and RM6378 is live in the future'
 
-      context 'and the framework is not RM3830 or RM6232' do
+      context 'and the framework is not RM6232 or RM6378' do
         it 'renders the unrecognised framework page with the right http status' do
           get :index, params: { framework: 'RM3826' }
 
@@ -64,18 +64,18 @@ RSpec.describe FacilitiesManagement::Admin::HomeController do
         end
       end
 
-      context 'and the framework is RM6232' do
+      context 'and the framework is RM6378' do
         it 'raises the MissingExactTemplate error' do
           expect do
-            get :index, params: { framework: 'RM6232' }
+            get :index, params: { framework: 'RM6378' }
           end.to raise_error(ActionController::MissingExactTemplate)
         end
       end
 
-      context 'and the framework is RM3830' do
+      context 'and the framework is RM6232' do
         it 'raises the MissingExactTemplate error' do
           expect do
-            get :index, params: { framework: 'RM3830' }
+            get :index, params: { framework: 'RM6232' }
           end.to raise_error(ActionController::MissingExactTemplate)
         end
       end